requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.1.0
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
exceptiongroup==1.1.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fonttools==4.40.0
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///C:/b/abs_49tt462f1q/croot/gitpython_1674662788764/work
humanize==4.6.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682176699712/work
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1685727936079/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
joblib==1.2.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1686775770267/work
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
nltk==3.8.1
numpy==1.24.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
plac==1.3.5
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1686403120743/work
pluggy==1.0.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyparsing==3.0.9
pytest==7.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2023.3
pywin32==305.1
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
regex==2023.6.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
spiral @ git+https://github.com/casics/spiral.git@0e75418cda7635f151fedff035b7b75799ccbe43
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
termcolor==2.3.0
tomli==2.0.1
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
tzdata==2023.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
